requirements.txt
Django>=3.0,<4.0
gunicorn>=20.0,<21.0
djangorestframework>=3.12